// src/interfaces/fighter_info.ts
export interface FighterInfo {
    Name: string;
    Nickname: string;
    Status: string;
    Age: string;
    Height: string;
    Weight: string;
    ArmReach: string;
    LegReach: string;
    FightingStyle: string;
    Division: string;
    PlaceOfBirth: string;
    TrainingCamp: string;
    OctagonDebut: string;
    ImageURL: string | undefined;
}